Camera module ready

+++ vec2.cpp
@@ -2,12 +2,12 @@
 #include <math.h>
 
 // Constructor
-Vec2::Vec2(float x, float y) 
+Vec2::Vec2(int x, int y) 
 	: x(x), y(y)
 {}
 
 // Length of the vector
 float Vec2::magnitude()
 {
-	return sqrt(x*x + y*y);
+	return sqrt((float)x*(float)x + (float)y*(float)y);
 }
